+++ nes/utility.py
@@ -55,7 +55,7 @@ def parse_nes_file(f):
 def _extract_header(f):
     _next = lambda: f.read(1)[0]
     i = 0
-    h = np.zeros(HEADER_SIZE, dtype='uint8')
+    h = np.zeros(HEADER_SIZE, dtype='int')
     for i in range(HEADER_SIZE):
         b = _next()
         h[i] = b
@@ -76,7 +76,7 @@ def _extract_chunk(f, size):
     """
     log.debug('Extracting chunk of size %s', size)
     _next = lambda: f.read(1)[0]
-    data = np.zeros(size, dtype='uint8')
+    data = np.zeros(size, dtype='int')
     for i in range(size):
         data[i] = _next()
     return data
